Сorrected remarks
