+++ Notebook/Client/Program.java
@@ -1,5 +1,5 @@
-package Notebook.Client;
-import Notebook.UI.Commands.App;
+package notebook.client;
+import notebook.core.mvp.views.UI.App;
 
 public class Program {
    public static void main(String[] args) {
