+++ Notebook/Core/MVP/Views/UI/commands/ShowAllCommand.java
@@ -1,6 +1,6 @@
-package Notebook.UI.Commands;
+package notebook.core.mvp.views.UI.commands;
 
-import Notebook.Core.MVP.Presenter;
+import notebook.core.mvp.Presenter;
 
 public class ShowAllCommand implements Command {
     private Presenter presenter;
