+++ Notebook/Core/MVP/Views/ViewRecord.java
@@ -1,4 +1,4 @@
-package Notebook.Core.MVP.Views;
+package notebook.core.mvp.views;
 
 public interface ViewRecord {
     String getTitle();
