+++ Notebook/DB/Confing.java
@@ -1,4 +1,4 @@
-package Notebook.DB;
+package notebook.db;
 
 public class Confing {
     public static String pathDb = "/Task05/HW_OOP_NB/Notebook/DB/data.json";
